Just update the glyph atlas when rendering nodes
Why was it not always like this? When the glyph atlas was too expensive
to render in one go (i.e. when I created distance fields on the client),
it was preferable to just show the empty glyphs (or actually, the
fast-rendered glyphs) and switch to the better ones over time. It never
stopped looking bad to me, so I dropped it in favor of just using the
fast versions.